+++ src/components/Icon.tsx
@@ -1,7 +1,7 @@
 type IconProps = {
   icon: string,
 }
-export function Icon({ icon } : IconProps) {
+export default function Icon({ icon } : IconProps) {
   const className = `bi bi-${icon} me-2`;
   return (
     <i className={className}></i>
