+++ src/hooks/form.tsx
@@ -1,0 +1,11 @@
+import { createFormHookContexts } from "@tanstack/react-form";
+import { createFormHook } from "@tanstack/react-form";
+
+export const { fieldContext, useFieldContext, formContext, useFormContext } = createFormHookContexts();
+
+export const { useAppForm, withForm, withFieldGroup } = createFormHook({
+  fieldComponents: {},
+  formComponents: {},
+  fieldContext,
+  formContext,
+});
