+++ src/routes/cases/components/AggressorFields.tsx
