+++ src/routes/cases/formValues.tsx
@@ -1,0 +1,15 @@
+export default {
+    victim: {
+        fullName: "",
+        age: "",
+        gender: "",
+        nationality: "",
+        isSexualWorker: false,
+        isMissingPerson: false,
+        isNativePeople: false,
+        isPregnant: false,
+        hasDisabillity: false,
+        occupation: "",
+        hasChildren: false,
+    },
+};
